Moved formatting function into utils.js
