+++ js/eventsList.js
@@ -1,7 +1,10 @@
 import { Event } from './modules.js';
+import { formatDateForCalendar } from './utils.js';
 
 // EventsList class definition
 export class EventsList {
+	formatDateForCalendar = formatDateForCalendar;
+
 	constructor() {
 		this.events = [];
 	}
@@ -155,62 +158,6 @@ export class EventsList {
 		ctx.fillText("📅", canvas.width / 2, canvas.height / 2);
 
 		return canvas.toDataURL();
-	}
-
-	// Helper method to format date and time for Google Calendar
-	formatDateForCalendar(date, startTime, endTime) {
-		// Google Calendar requires the date-time in the format 20211001T100000Z/20211001T110000Z
-		// console.log('date', date);
-		// console.log('startTime', startTime);
-		// console.log('endTime', endTime);
-
-		let dateFormatted = date.replace('-', '').replace('-', ''); // Remove dashes from the date string
-		let startTimeFormatted = startTime;
-		let endTimeFormatted = endTime;
-
-		// console.log('dateFormatted', dateFormatted);
-
-		// If PM, add 12 hours to the start time, upper or lower case
-		if (startTime.toLowerCase().indexOf("p.m.") != -1 && startTime.split(':')[0] !== '12') {
-			const [hours, minutes] = startTime.split(':');
-			startTimeFormatted = `${parseInt(hours) + 12}:${minutes}`;
-		}
-		// If AM, set hours to 0 if it's 12
-		if (startTime.toLowerCase().indexOf("a.m.") != -1 && startTime.split(':')[0] === '12') {
-			const [hours, minutes] = startTime.split(':');
-			startTimeFormatted = `00:${minutes}`;
-		}
-		// If PM, add 12 hours to the end time
-		if (endTime.toLowerCase().indexOf("p.m.") != -1 && endTime.split(':')[0] !== '12') {
-			const [hours, minutes] = endTime.split(':');
-			endTimeFormatted = `${parseInt(hours) + 12}:${minutes}`;
-		}
-		// If AM, set hours to 0 if it's 12
-		if (endTime.toLowerCase().indexOf("a.m.") != -1 && endTime.split(':')[0] === '12') {
-			const [hours, minutes] = endTime.split(':');
-			endTimeFormatted = `00:${minutes}`;
-		}
-		// Trim whitespace
-		startTimeFormatted = startTimeFormatted.trim();
-		endTimeFormatted = endTimeFormatted.trim();
-		// Remove a.m. and p.m. from the time strings
-		startTimeFormatted = startTimeFormatted.toLowerCase().replace(/a\.m\./i, '').replace(/p\.m\./i, '');
-		endTimeFormatted = endTimeFormatted.toLowerCase().replace(/a\.m\./i, '').replace(/p\.m\./i, '');
-		// console.log('startTime', startTime);
-		// Remove colons from the time strings
-		startTimeFormatted = startTimeFormatted.replace(/:/g, '');
-		endTimeFormatted = endTimeFormatted.replace(/:/g, '');
-
-		// Format the start and end times
-		const startDateFormatted = dateFormatted + 'T' + startTimeFormatted.replace(' ', '') + '00'; // Assuming UTC+5:00
-		const endDateFormatted = dateFormatted + 'T' + endTimeFormatted.replace(' ', '') + '00'; // Assuming UTC+5:00
-
-		// console.log('startDateFormatted', startDateFormatted);
-		// console.log('endDateFormatted', endDateFormatted);
-
-		const dateRangeString = `${startDateFormatted}/${endDateFormatted}`;
-		// console.log(dateRangeString);
-		return dateRangeString;
 	}
 
 	// Helper function to check if a date string is in ISO 8601 format
